DEVELOPMENT: first feature test
Would not normally commit while failing a test, but this is a feature test, and trying to be as transaparent
as possible. Intention of first test is to be pretty comprehensive, will (via smaller unit tests) create the
Item class, Checkout class, scan method, total method and have structure within the item class with attributes.

Quite a large feature, but going to write seperate unit tests to drive its development incrementally... i hope!