ADD: ITEM PRODUCT CODE
Dont like this test as its kind of circular and not testing behaviour, it's testing state, so its not ideal, but it'll do for now as i have a feature
test checking behaviour and this is lower level and establishing the initialization of the Item object. It'll do, moving on.