ADD: checkout, item tests and code.
Tests are still failing, committing just for clarity sake, progressing through the RED, GREEN, REFACTOR process... feature test first had failing item
so wrote simple test for that, code for that which then was passing, so had feature test failing and one passing test, so wrote test for checkout test,
had an error as usign same name for Item in the test setup, so changed it (will move to helpers with a refactor). So have failling checkout test, and
currently have a feature test and a unit test failing for the same reason - no scan method has yet been created on the checkout class... so back to making
tests pass.