REFACTOR: checkout
So the tests were passing as i have performed the minimum to pass them, but still i was unhappy with the quality as basically checkout
was hard coded.. indicating i needed more tests, so a couple more tests were written, expecting different outputs, including if nothing
had been added the the checkout basket.

Also thought that it would be a good idea to start using let as a helper to demonstrate DRY code.