STYLING: standard gem, private helper, logic on multiple objects in basket
Generally a cosmetic update, added some logic to the total method, which got too large and therefore extracted that into a private method, could probably further refine that into another 'calculator' type class... later.

Have wanted to use the standard gem for a while so took the opportunity to have a read and use that via $ standardrb --fix to auto fix styling on the project... nice. It does however use double quotes not single.

Also had an issue with multi items and string interpolation on the print of total, now fixed. Have x'd out the total price test for now... moving onto that in a moment.