REFACTOR / DEV: pulled out calculator class and changed tests a bit.
Breaking the code down and playing a bit, inserted basket calculator as a module as feel
like it is the right thing to do. That then gives me access to list items method, which
is the first of two tasks that need to be done pre any discounts. Next would be to calculate
the total price of objects in the basket... again have failing tests, at the moment, but trying
to clarify steps while i fiddle through some refactoring. Trying to break the classes and bit
and encapsulate better. Tests needed extending, have added more output with the total method
with the puts adding a new line and total values on a new line of the total method. Next is to
get the calculate price method right.. and the formatting of the number into a sterling styled
figure, i found a gem called money... but that is apparently being depreciated, so after
downloading I removed it. Onwards ho!! :wq