ADD: calculate price method on the basket calculator... complete
Test is passing, as i thought the rounding was the issue so change to a float, that exposed a
single quoted test, which didnt return the new lines, therefore fixed that.... test are now
passing!

So thats red... tests, green... code passing, refactor... fix/improve things... going to now
spend some time refactoring.