Tidied code
Being a plonker i left the workings of the calculator in the checkout... removed that and added a
multiple items into basket test as appear to have missed that. Tests are checking behaviour and outputs
so think that they are ok at the moment, not as concerned about the modules as main focus is on the
public interface of the code.