ADD: promotion.... starting the 'actual' logic bits
so with the initialization of the checkout there is the addition of a promotion,
if i had free option to do so i think that a method to add the promotion
would be better, but i guess a new checkout is created at each attendance at the
till so its ok. Feels like the promotion may change during the checkout process,
so following the guidance as that is what was asked for.

First couple of tests to create the promotion object and set a default benefit
given of zero. Have decided that i want the promotion to look at the basket and
provide a reduction to the total price... and either adjust total method or
refactor calculate price so that it just returns a number (probably better) and
that number is used and a number that promotion uses is used in another private
method in a 'printer' that prints the formatted string with the price - whether
that be a detault no promotion price, or a price with the promotion on it.
Hope that makes sense! i.e. better encapsulate the methods, breking it down a bit
more.