REFACTOR: calculating...
getting a bit messy, so created a module for promotion calculator, and promotion just has attributes, for
price over and basket discount... detaulted to zero at the momement so existing tests pass. At the intialzation the
checkout and its param promotion can have a specific named argument added to it... so the promotion can be varied
when its created with different criteria.. felt like it'd be good to encapsulate the calculation away from the
promotion element itself.