ADD: promotion!
using params on the promotion object you can create custom promotions, and the promotion
calculator can be used to provide a public benefit variable which is reduced from the
basket price... resulting in a promotion working.. all that is left is to set logic in the
promotion calculator to return the benefit under different conditions. Want to refeactor somewhat
and to tidy up.